refactor(Command): Rename V2 Command to Command
